# Add expander functionality