@kraih 2016-04-23
Pretty sure this breaks JSON decoding for non-blocking queries.