@chy-causer 2016-04-23
I have just spotted a mistake in the bench.pl script. Corrected in the gist, but will need to rerun to regenerate numbers. It should only affect absolute numbers, not ratios, but I'd like to be 100% confident in my results.