@chy-causer 2016-04-23
If this breaks non-blocking queries, then there are missing tests as all existing tests pass.