@kraih 2016-04-23
No, you forgot to add new tests. Until now there was no difference between blocking and non-blocking queries for JSON decoding.