@kraih 2016-04-23
Changing from `Mojo::JSON` to `JSON::XS` is a non-issue btw., since `Mojo::JSON::MaybeXS` exists.