@kraih 2016-04-24
I think a pretty strong argument against this functionality might be that the way `DBD::Pg` handles different data types is very sketchy. Types like `_json` and `_jsonb` are not even mentioned in the documentation, `bigint` seems like something `DBD::Pg` might start handling itself at any time.