@kraih 2016-04-24
Like i said on IRC before, the patch for this feature has to make the expand code cleaner to be worth it for me, and i just don't think that's the case here. So i'm afraid i'll have to give it a 👎  vote.