@jberger 2016-04-24
I haven't given this a ton of review admittedly, since I've been busy at QAH. I do worry that it seems quite complicated, especially given that most of this functionality can be achieved by coercion in the db query itself. Some performance gain is nice but perhaps not worth the additional complexity. (Note that this is a very common consideration in Mojo-land and not just a comment on this individual PR). As such I'm 👎 at least until the complexity can be minimized.