@chy-causer 2016-04-24
Thank you for looking past the code sloppiness to get to what I am trying to achieve. I do not have time at the moment to prepare a full response so this is an email to say I share some concerns raised, and there are some that I feel can be defended. I will try to reply point by point within the next 24 hours.