@jberger 2016-04-24
@chy-causer note that I've updated my comment a little (to clarify). The sentence is now:

"Some performance gain of things like a bool expander is nice but perhaps not worth the additional complexity (especially given that other existing expanders see a performance hit). "