@chy-causer 2016-04-24
If I have not explicitly addressed any points raised in previous comments, please take that as me not having any response of note.

# Benchmarking revisited

The only system I have available to do benchmarking is hosted on a fairly busy VI, so the results are rather noisy. Topping that, there was a fairly howling error in the original benchmarking script and the results were skewed. Here is a new benchmark, with the numbers of iterations and rows boosted, and  the script niced to -10 to reduce noise further. I have made no modifications suggested in the comments about wasteful calls.

|Benchmark|Before|After|
|---------|------|-----|
|expand->hashes  |52.4/s | 52.8/s|
|expand->hash    |36.2/s | 36.3/s|
|hashes          |349.1/s |353.3/s |
|bool::json      |95.5/s | 102.2/s|
|Extra expander  | -   | 102.4/s|
|bool expander   | -   | 193.8/s|

Rows: 2000
Iterations: 5000 (10000 for raw hashes)
Before: 290.6s
After: 353.7s

As an aside, I now have much more sympathy for the original author of _expand as I can now see that much of its design was to boost performance.

You are welcome to rerun these benchmarks yourself to verify.

https://gist.github.com/chy-causer/66b15f14b16b06d3d9da0e36cd36ca04

# _column_types and _columns_by_types being needless

They would be required for the following functionality:

    $db->query($QUERY)->expand(column_name => \&expander);

You're not mind-readers, I didn't explain it, and I was getting ahead of myself. I will not pull something like that again.

# Of DBD::Pg return types

I wholeheartedly agree that return types are sketchy and unintuitive at times. DBD::Pg's handling can also be fairly inconsistent. Consider the two stanzas and see if you can guess what they return:

    $pg->db->query(
      'select array_agg(distinct state) from minion_jobs'
    )->array;

    $pg->db->query(
      'select array_agg(distinct state)::text[] from minion_jobs'
    )->array;

However, on the flip side, the types that do exist, undocumented though they are, have been stable for a number of years now.

https://github.com/autarch/DBD-Pg/commits/master/types.c

`int8` is the internal representation of bigint by PostgreSQL and is most likely here to stay. Most other important types are fairly intuitive and I still think there is benefit to be had in expanding them (xml, booleans, timestamps and the one I actually care about, cidr.)

The others like _json I can happily ignore. Arrays of json aren't expanded currently, so I don't see any reason to start.

As a compromise, Mojo::Pg could limit allowed expansion types to a well defined subset of DBD::Pg types (even aliasing the non-intuitive ones like int8.)

# Bigint may be handled by DBD::Pg any time

How would it handle it? By returning a Math::BigInt object (the only core module I can think of that would be appropriate)? For arguments sake if they do do that, then Math::BigInt can take Math::BigInt objects as an argument, so nothing will break should that happen. I cannot speak for the developers of DBD::Pg, but I think the likelihood of that happening is small.

# What's proposed can be achieved by coercion within the database itself

That's absolutely not the case.

    $db->query(
      "select '<hello>world</hello>'::xml"
    )->expand->array->[0]->at('hello');

This cannot be done via coercion within the database.

Similarly, even worse, something that you think works, but gives the wrong answer:

    $db->query('SELECT 9223372036854775807::bigint')->array->[0] / 10;

# The change breaks non-blocking

I'm afraid I cannot write a test that works on the original code but not the new one so I will need some assistance here.

# Speed increase of boolean expansion not worth the complexity

Complexity for whom? If it's the Mojo::Pg codebase then I don't have an answer for you other than a promise to go back and look at the code. If it's the end user, then I would argue that

    $db->('SELECT true')->expand->array;

Is much less complex than

    $db->('SELECT json_agg(true)')->expand->array->[0];

Obviously you wouldn't do these queries in real life, but hopefully you get the picture.

On top of that, the new benchmark shows around 100% performance improvement for the former over the latter (please do verify this for yourself though.)