@jberger 2016-04-24
Just about to go to bed here in England as the QAH wraps up and I fly home. Just before I do go, yes I did mean for the Mojo::Pg codebase. While we certainly care about the end users, we do value the cleanliness of the codebase too. This is not to say that the idea is dead, just that it would need tightening up before it could really be considered. As you have looked through the Mojo code I'm sure you've noticed the style :-P

One other quick thing, this isn't very easy to immediately understand:

    $db->('SELECT json_agg(true)')->expand->array->[0];

but I think this reasonably is

    $db->('SELECT to_json(true)')->expand->array;

Perhaps you would like to join #mojo on irc.perl.org to discuss further? PRs are nice for reviewing code but to discuss things like api and utility irc is really nice. I'm usually on and around (travel notwithstanding). My nick is jberger. @kraih is sri.