@chy-causer 2016-04-25
I appreciate you want to keep the code clean. As I said earlier, this is my
first PR, pretty much for any project, not just Mojo::Pg, so I'm well aware
there is much tightening up that can be done and I'm very grateful you and
@kraih took the time to read the code and discuss its merits, rather than
dismiss it out of hand.

I am on IRC as well, but am only able to respond during UK work hours. My
nick is CHYC.